Revert "Bugfixing (#5376)"
This reverts commit 6ae1f7bd4bcfa0349a2b1e0cac6b45726ae82ea9.
