+++ decidim-assemblies/config/locales/en.yml
@@ -264,7 +264,7 @@ en:
         others: Others
         public: Public
       filter:
-        all: All types
+        all: All types of assemblies
         commission: Commission
         consultative_advisory: Consultative/Advisory
         executive: Executive
