+++ decidim-budgets/app/views/decidim/budgets/projects/_filters.html.erb
@@ -19,4 +19,6 @@
   <% if current_component.categories.any? %>
     <%= form.categories_select :category_id, current_component.categories, legend_title: t(".category"), disable_parents: false, label: false, prompt: t(".category_prompt") %>
   <% end %>
+
+  <%= form.hidden_field :random_seed %>
 <% end %>
