+++ decidim-core/app/cells/decidim/activities_cell.rb
@@ -13,6 +13,13 @@ module Decidim
     def show
       return if activities.blank?
 
+      activities.map do |activity|
+        activity.organization_lazy
+        activity.resource_lazy
+        activity.participatory_space_lazy
+        activity.component_lazy
+      end
+
       render
     end
 
@@ -27,19 +34,7 @@ module Decidim
     end
 
     def activities
-      @activities ||= last_activities.select do |activity|
-        activity.visible_for?(current_user)
-      end
-    end
-
-    def last_activities
-      @last_activities ||= model.map do |activity|
-        activity.organization_lazy
-        activity.resource_lazy
-        activity.participatory_space_lazy
-        activity.component_lazy
-        activity
-      end
+      model
     end
   end
 end
