+++ decidim-core/app/models/decidim/action_log.rb
@@ -128,17 +128,5 @@ module Decidim
         scope.each { |relation| loader.call(relation.id, relation) }
       end
     end
-
-    # Whether this activity or log is visible for a given user (can also be nil)
-    #
-    # Returns a True/False.
-    def visible_for?(user)
-      return false if resource_lazy.blank?
-      return false if participatory_space_lazy.blank?
-      return false if resource_lazy.respond_to?(:hidden?) && resource_lazy.hidden?
-      return false if resource_lazy.respond_to?(:can_participate?) && !resource_lazy.can_participate?(user)
-
-      true
-    end
   end
 end
