+++ decidim-generators/exe/decidim
@@ -2,8 +2,8 @@
 # frozen_string_literal: true
 
 if File.exist?(File.expand_path("../../.git", __dir__))
-  ENV["RUBYOPT"] = ENV["RUBYOPT"].to_s.sub("-rsimplecov ", "")
-  ENV["RUBYLIB"] = ENV["RUBYLIB"].to_s.sub("#{File.expand_path("..", __dir__)}/lib:", "")
+  ENV["RUBYOPT"] = ENV["RUBYOPT"].sub("-rsimplecov ", "")
+  ENV["RUBYLIB"] = ENV["RUBYLIB"].sub("#{File.expand_path("..", __dir__)}/lib:", "")
 
   gem "decidim-core"
 end
@@ -16,11 +16,9 @@ when "--component"
 
   Decidim::Generators::ComponentGenerator.start
 when "--version"
-  require "decidim/core/version"
-  puts Decidim::Core.version
+  puts Decidim.version
 when "--help"
-  require "decidim/core/version"
-  puts "Decidim #{Decidim::Core.version}
+  puts "Decidim #{Decidim.version}
 Syntax: decidim [options] [App or Module name]
 https://decidim.org
