+++ decidim-generators/lib/decidim/generators/component_generator.rb
@@ -1,7 +1,6 @@
 #!/usr/bin/env ruby
 # frozen_string_literal: true
 
-require "bundler"
 require "thor"
 require "active_support/inflector"
 require "decidim/core/version"
