+++ decidim-initiatives/app/views/decidim/initiatives/admin/initiatives/_form.html.erb
@@ -29,7 +29,7 @@
                         {
                           disabled: !@form.signature_type_updatable?,
                           "data-scope-selector": "initiative_decidim_scope_id",
-                          "data-scope-id": form.object.decidim_scope_id,
+                          "data-scope-id": current_initiative.scoped_type.id,
                           "data-scope-search-url": decidim_initiatives.initiative_type_scopes_search_url,
                           "data-signature-types-selector": "initiative_signature_type",
                           "data-signature-type": current_initiative.signature_type,
