+++ decidim-initiatives/app/views/decidim/initiatives/initiatives/_initiatives.html.erb
@@ -11,4 +11,4 @@
 <div class="row small-up-1 medium-up-2 card-grid">
   <%= render initiatives %>
 </div>
-<%= decidim_paginate initiatives %>
+<%= decidim_paginate initiatives, random_seed: random_seed %>
