+++ decidim-proposals/app/views/decidim/proposals/collaborative_drafts/_collaborative_drafts.html.erb
@@ -13,4 +13,4 @@
     <%= card_for draft, context: { label: true } %>
   <% end %>
 </div>
-<%= decidim_paginate @collaborative_drafts %>
+<%= decidim_paginate @collaborative_drafts, random_seed: random_seed %>
