+++ decidim-proposals/app/views/decidim/proposals/proposals/_proposals.html.erb
@@ -11,4 +11,4 @@
 <div class="row small-up-1 medium-up-2 card-grid">
   <%= render @proposals %>
 </div>
-<%= decidim_paginate @proposals %>
+<%= decidim_paginate @proposals, random_seed: random_seed %>
