+++ decidim-system/app/views/decidim/system/organizations/show.html.erb
@@ -7,4 +7,5 @@
 <div class="actions">
   <hr>
   <%= link_to t("decidim.system.actions.edit"), ["edit", @organization] %>
+  <%= link_to t("decidim.system.actions.destroy"), @organization, method: :delete, class: "alert button", data: { confirm: t("decidim.system.actions.confirm_destroy") } %>
 </div>
